Move library flags to the end of the command line.
I know that on some platforms (e.g., Windows) this has been necessary in
the past. I don't know if it still affects certain platforms or if it
would affect any native UNIX-likes, but it's worth a shot if it might
help other people compile. It shouldn't affect sane platforms either
way.